requirements.txt
# 🚀 Agents Flow - Multi-Agent Recommendation System
# Core dependencies for the recommendation system
# Updated for Python 3.13 compatibility

# =============================================================================
# Web Framework & API
# =============================================================================
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# =============================================================================
# Data Processing & Analysis
# =============================================================================
pandas>=2.2.0
numpy>=1.26.0

# =============================================================================
# LangChain & LangGraph Framework
# =============================================================================
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.3.0
langgraph>=0.2.0

# =============================================================================
# AI/ML & Embeddings
# =============================================================================
openai>=1.3.0
anthropic>=0.57.0

# =============================================================================
# Data Validation & Serialization
# =============================================================================
pydantic>=2.5.0
pydantic-settings>=2.0.0
typing-extensions>=4.8.0

# =============================================================================
# Environment & Configuration
# =============================================================================
python-dotenv>=1.0.0

# =============================================================================
# Fuzzy Matching & Text Processing
# =============================================================================
rapidfuzz>=3.6.0

# =============================================================================
# Development & Testing
# =============================================================================
IPython>=8.17.0
requests>=2.31.0

# =============================================================================
# Optional: UI Frameworks (for future use)
# =============================================================================
# streamlit>=1.29.0
# gradio>=4.44.0

# =============================================================================
# Development Tools (optional)
# =============================================================================
# pytest>=7.4.0
# black>=23.11.0
# flake8>=6.1.0
# mypy>=1.7.0
